+++ builds/appchan-x.user.js
@@ -256,7 +256,7 @@
         'Mark Quotes of You': [true, 'Add \'(You)\' to quotes linking to your posts.'],
         'Highlight Posts Quoting You': [false, 'Highlights any posts that contain a quote to your post.', 1],
         'Highlight Own Posts': [false, 'Highlights own posts if Quote Markers are enabled.', 1],
-        'Double Beep': [false, 'If Beep is enabled, then beeps twice to notify of a post quoting you.', 1],
+        'Double Beep': [false, 'Beeps twice to notify you of a post made that is quoting you.', 1],
         'Mark OP Quotes': [true, 'Add \'(OP)\' to OP quotes.'],
         'Mark Cross-thread Quotes': [true, 'Add \'(Cross-thread)\' to cross-threads quotes.'],
         'Quote Threading': [false, 'Thread conversations']
@@ -8458,7 +8458,7 @@
           markers.push('You');
         }
         $.addClass(post.nodes.root, 'quotesYou');
-        if (Conf['Beep'] && Conf['Double Beep']) {
+        if (Conf['Double Beep']) {
           QuoteMarkers.beep = true;
         }
       }
@@ -14670,7 +14670,7 @@
       ThreadUpdater.outdateCount = 0;
       ThreadUpdater.lastPost = posts[count - 1].ID;
       Post.callbacks.execute(posts);
-      if (Conf['Beep'] && d.hidden && Unread.posts && !Unread.posts.length) {
+      if ((Conf['Beep'] || Conf['Double Beep']) && d.hidden && Unread.posts && !Unread.posts.length) {
         if (!ThreadUpdater.audio) {
           ThreadUpdater.audio = $.el('audio', {
             src: ThreadUpdater.beep,
@@ -14682,7 +14682,9 @@
             }
           });
         }
-        ThreadUpdater.audio.play();
+        if (QuoteMarkers.beep || Conf['Beep']) {
+          ThreadUpdater.audio.play();
+        }
       }
       scroll = Conf['Auto Scroll'] && ThreadUpdater.scrollBG() && Header.getBottomOf(ThreadUpdater.root) > -75;
       for (_j = 0, _len1 = posts.length; _j < _len1; _j++) {